Allow more memory to tests, except streaming tests
Motivation:

We currently allow only 128 MiB of heap for tests, which is because we
do not want the streaming tests in `HttpServerTest` take too long
streaming `2 * maxMemory` bytes.

We don't really need to restrict the heap size for other tests. Small
heap size increases the test time for some complex tests that demands
more memory.

Modifications:

- Extract the streaming tests in `HttpServerTest` into
  `HttpServerStreamingTest`
- Exclude `HttpServerStreamingTest` from the `test` task
- Add `testStreaming` task which runs `HttpServerStreamingTest` only
- Allow only 64 MiB to `testStreaming` while allowing 512 MiB to other
  tests

Result:

- Streaming tests end sooner.
- Reduced memory pressure while running non-streaming tests
