+++ core/src/test/java/com/linecorp/armeria/server/HttpServerTest.java
@@ -40,10 +40,8 @@ import java.util.concurrent.BlockingQueue;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.LinkedBlockingQueue;
-import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicInteger;
-import java.util.concurrent.atomic.AtomicReference;
 import java.util.function.Function;
 import java.util.zip.GZIPInputStream;
 import java.util.zip.InflaterInputStream;
@@ -60,11 +58,8 @@ import org.junit.runners.Parameterized;
 import org.junit.runners.Parameterized.Parameters;
 import org.reactivestreams.Subscriber;
 import org.reactivestreams.Subscription;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import com.google.common.base.Strings;
-import com.google.common.base.Throwables;
 import com.google.common.collect.ImmutableList;
 import com.google.common.io.ByteStreams;
 
@@ -88,10 +83,8 @@ import com.linecorp.armeria.common.MediaType;
 import com.linecorp.armeria.common.SessionProtocol;
 import com.linecorp.armeria.common.logging.RequestLog;
 import com.linecorp.armeria.common.logging.RequestLogAvailability;
-import com.linecorp.armeria.common.stream.StreamWriter;
 import com.linecorp.armeria.common.util.EventLoopGroups;
 import com.linecorp.armeria.common.util.Exceptions;
-import com.linecorp.armeria.internal.InboundTrafficController;
 import com.linecorp.armeria.internal.PathAndQuery;
 import com.linecorp.armeria.server.encoding.HttpEncodingService;
 import com.linecorp.armeria.testing.server.ServerRule;
@@ -103,13 +96,9 @@ import io.netty.channel.EventLoopGroup;
 import io.netty.handler.ssl.util.InsecureTrustManagerFactory;
 import io.netty.util.AsciiString;
 import io.netty.util.NetUtil;
-import io.netty.util.ResourceLeakDetector;
-import io.netty.util.ResourceLeakDetector.Level;
-import io.netty.util.concurrent.GlobalEventExecutor;
 
 @RunWith(Parameterized.class)
 public class HttpServerTest {
-    private static final Logger logger = LoggerFactory.getLogger(HttpServerTest.class);
 
     private static final EventLoopGroup workerGroup = EventLoopGroups.newEventLoopGroup(1);
 
@@ -120,11 +109,6 @@ public class HttpServerTest {
             .build();
 
     private static final long MAX_CONTENT_LENGTH = 65536;
-    // Stream as much as twice of the heap. Stream less to avoid OOME when leak detection is enabled.
-    private static final long STREAMING_CONTENT_LENGTH =
-            Runtime.getRuntime().maxMemory() * 2 / (ResourceLeakDetector.getLevel() == Level.PARANOID ? 8 : 1);
-    private static final int STREAMING_CONTENT_CHUNK_LENGTH =
-            (int) Math.min(Integer.MAX_VALUE, STREAMING_CONTENT_LENGTH / 8);
 
     @Parameters(name = "{index}: {0}")
     public static Collection<SessionProtocol> parameters() {
@@ -305,22 +289,6 @@ public class HttpServerTest {
                             res.close();
                         }
                     });
-                    return res;
-                }
-            });
-
-            sb.service("/count", new CountingService(false));
-            sb.service("/slow_count", new CountingService(true));
-
-            sb.serviceUnder("/zeroes", new AbstractHttpService() {
-                @Override
-                protected HttpResponse doGet(ServiceRequestContext ctx, HttpRequest req) {
-                    final long length = Long.parseLong(ctx.mappedPath().substring(1));
-                    final HttpResponseWriter res = HttpResponse.streaming();
-                    res.write(HttpHeaders.of(HttpStatus.OK)
-                                         .setLong(HttpHeaderNames.CONTENT_LENGTH, length));
-
-                    stream(res, length, STREAMING_CONTENT_CHUNK_LENGTH);
                     return res;
                 }
             });
@@ -592,41 +560,11 @@ public class HttpServerTest {
     }
 
     @Test(timeout = 10000)
-    public void testTooLargeContent() throws Exception {
-        clientWriteTimeoutMillis = 0L;
-        final HttpRequestWriter req = HttpRequest.streaming(HttpMethod.POST, "/count");
-        final CompletableFuture<AggregatedHttpMessage> f = client().execute(req).aggregate();
-
-        stream(req, MAX_CONTENT_LENGTH + 1, 1024);
-
-        final AggregatedHttpMessage res = f.get();
-
-        assertThat(res.status()).isEqualTo(HttpStatus.REQUEST_ENTITY_TOO_LARGE);
-        assertThat(res.headers().contentType()).isEqualTo(MediaType.PLAIN_TEXT_UTF_8);
-        assertThat(res.content().toStringUtf8()).isEqualTo("413 Request Entity Too Large");
-    }
-
-    @Test(timeout = 10000)
     public void testTooLargeContentToNonExistentService() throws Exception {
         final byte[] content = new byte[(int) MAX_CONTENT_LENGTH + 1];
         final AggregatedHttpMessage res = client().post("/non-existent", content).aggregate().get();
         assertThat(res.headers().status()).isEqualTo(HttpStatus.NOT_FOUND);
         assertThat(res.content().toStringUtf8()).isEqualTo("404 Not Found");
-    }
-
-    @Test(timeout = 60000)
-    public void testStreamingRequest() throws Exception {
-        runStreamingRequestTest("/count");
-    }
-
-    @Test(timeout = 120000)
-    public void testStreamingRequestWithSlowService() throws Exception {
-        final int oldNumDeferredReads = InboundTrafficController.numDeferredReads();
-        runStreamingRequestTest("/slow_count");
-        // The connection's inbound traffic must be suspended due to overwhelming traffic from client.
-        // If the number of deferred reads did not increase and the testStreaming() above did not fail,
-        // it probably means the client failed to produce enough amount of traffic.
-        assertThat(InboundTrafficController.numDeferredReads()).isGreaterThan(oldNumDeferredReads);
     }
 
     @Test(timeout = 10000)
@@ -820,50 +758,6 @@ public class HttpServerTest {
         }
     }
 
-    private void runStreamingRequestTest(String path) throws InterruptedException, ExecutionException {
-        // Disable timeouts and length limits so that test does not fail due to slow transfer.
-        clientWriteTimeoutMillis = 0;
-        clientResponseTimeoutMillis = 0;
-        serverRequestTimeoutMillis = 0;
-        serverMaxRequestLength = 0;
-
-        final HttpRequestWriter req = HttpRequest.streaming(HttpMethod.POST, path);
-        final CompletableFuture<AggregatedHttpMessage> f = client().execute(req).aggregate();
-
-        // Stream a large of the max memory.
-        // This test will fail if the implementation keep the whole content in memory.
-        final long expectedContentLength = STREAMING_CONTENT_LENGTH;
-
-        try {
-            stream(req, expectedContentLength, STREAMING_CONTENT_CHUNK_LENGTH);
-
-            final AggregatedHttpMessage res = f.get();
-
-            assertThat(res.status()).isEqualTo(HttpStatus.OK);
-            assertThat(res.headers().contentType()).isEqualTo(MediaType.PLAIN_TEXT_UTF_8);
-            assertThat(res.content().toStringUtf8()).isEqualTo(
-                    String.valueOf(expectedContentLength));
-        } finally {
-            // Make sure the stream is closed even when this test fails due to timeout.
-            req.close();
-        }
-    }
-
-    @Test(timeout = 60000)
-    public void testStreamingResponse() throws Exception {
-        runStreamingResponseTest(false);
-    }
-
-    @Test(timeout = 120000)
-    public void testStreamingResponseWithSlowClient() throws Exception {
-        final int oldNumDeferredReads = InboundTrafficController.numDeferredReads();
-        runStreamingResponseTest(true);
-        // The connection's inbound traffic must be suspended due to overwhelming traffic from client.
-        // If the number of deferred reads did not increase and the testStreaming() above did not fail,
-        // it probably means the client failed to produce enough amount of traffic.
-        assertThat(InboundTrafficController.numDeferredReads()).isGreaterThan(oldNumDeferredReads);
-    }
-
     @Test(timeout = 30000)
     public void testStreamRequestLongerThanTimeout() throws Exception {
         // Disable timeouts and length limits so that test does not fail due to slow transfer.
@@ -885,42 +779,6 @@ public class HttpServerTest {
         request.write(HttpData.ofUtf8("e"));
         request.close();
         assertThat(response.aggregate().get().content().toStringUtf8()).isEqualTo("abcde");
-    }
-
-    private void runStreamingResponseTest(boolean slowClient) throws InterruptedException, ExecutionException {
-        // Disable timeouts and length limits so that test does not fail due to slow transfer.
-        clientWriteTimeoutMillis = 0;
-        clientResponseTimeoutMillis = 0;
-        clientMaxResponseLength = 0;
-        serverRequestTimeoutMillis = 0;
-
-        final HttpResponse res = client().get("/zeroes/" + STREAMING_CONTENT_LENGTH);
-        final AtomicReference<HttpStatus> status = new AtomicReference<>();
-
-        final StreamConsumer consumer = new StreamConsumer(GlobalEventExecutor.INSTANCE, slowClient) {
-
-            @Override
-            public void onNext(HttpObject obj) {
-                if (obj instanceof HttpHeaders) {
-                    status.compareAndSet(null, ((HttpHeaders) obj).status());
-                }
-                super.onNext(obj);
-            }
-
-            @Override
-            public void onError(Throwable cause) {
-                // Will be notified via the 'awaitClose().get()' below.
-            }
-
-            @Override
-            public void onComplete() {}
-        };
-
-        res.subscribe(consumer);
-
-        res.completionFuture().get();
-        assertThat(status.get()).isEqualTo(HttpStatus.OK);
-        assertThat(consumer.numReceivedBytes()).isEqualTo(STREAMING_CONTENT_LENGTH);
     }
 
     @Test(timeout = 10000)
@@ -979,27 +837,6 @@ public class HttpServerTest {
                                        .contains("/cached-paths/hoge");
     }
 
-    private static void stream(StreamWriter<HttpObject> writer, long size, int chunkSize) {
-        if (!writer.tryWrite(HttpData.of(new byte[chunkSize]))) {
-            return;
-        }
-
-        final long remaining = size - chunkSize;
-        logger.info("{} bytes remaining", remaining);
-
-        if (remaining == 0) {
-            writer.close();
-            return;
-        }
-
-        writer.onDemand(() -> stream(writer, remaining, (int) Math.min(remaining, chunkSize)))
-              .exceptionally(cause -> {
-                  logger.warn("Unexpected exception:", cause);
-                  writer.close(cause);
-                  return null;
-              });
-    }
-
     private HttpClient client() {
         if (client != null) {
             return client;
@@ -1019,84 +856,5 @@ public class HttpServerTest {
                 });
 
         return client = builder.build();
-    }
-
-    private static class CountingService extends AbstractHttpService {
-
-        private final boolean slow;
-
-        CountingService(boolean slow) {
-            this.slow = slow;
-        }
-
-        @Override
-        protected HttpResponse doPost(ServiceRequestContext ctx, HttpRequest req) {
-            final CompletableFuture<HttpResponse> responseFuture = new CompletableFuture<>();
-            final HttpResponse res = HttpResponse.from(responseFuture);
-            req.subscribe(new StreamConsumer(ctx.eventLoop(), slow) {
-                @Override
-                public void onError(Throwable cause) {
-                    responseFuture.complete(
-                            HttpResponse.of(
-                                    HttpStatus.INTERNAL_SERVER_ERROR,
-                                    MediaType.PLAIN_TEXT_UTF_8,
-                                    Throwables.getStackTraceAsString(cause)));
-                }
-
-                @Override
-                public void onComplete() {
-                    responseFuture.complete(
-                            HttpResponse.of(
-                                    HttpStatus.OK, MediaType.PLAIN_TEXT_UTF_8, "%d", numReceivedBytes()));
-                }
-            });
-            return res;
-        }
-    }
-
-    private abstract static class StreamConsumer implements Subscriber<HttpObject> {
-        private final ScheduledExecutorService executor;
-        private final boolean slow;
-
-        private Subscription subscription;
-        private long numReceivedBytes;
-        private int numReceivedChunks;
-
-        protected StreamConsumer(ScheduledExecutorService executor, boolean slow) {
-            this.executor = executor;
-            this.slow = slow;
-        }
-
-        protected long numReceivedBytes() {
-            return numReceivedBytes;
-        }
-
-        @Override
-        public void onSubscribe(Subscription subscription) {
-            this.subscription = subscription;
-            subscription.request(1);
-        }
-
-        @Override
-        public void onNext(HttpObject obj) {
-            if (obj instanceof HttpData) {
-                numReceivedBytes += ((HttpData) obj).length();
-            }
-
-            if (numReceivedBytes >= (numReceivedChunks + 1L) * STREAMING_CONTENT_CHUNK_LENGTH) {
-                numReceivedChunks++;
-
-                if (slow) {
-                    // Add 1 second delay for every chunk received.
-                    executor.schedule(() -> subscription.request(1), 1, TimeUnit.SECONDS);
-                } else {
-                    subscription.request(1);
-                }
-
-                logger.debug("{} bytes received", numReceivedBytes);
-            } else {
-                subscription.request(1);
-            }
-        }
     }
 }
